implemented test code and modified service method signature
# Purpose
- 정말 단순한 파라미터(게시글 ID 등)만 요구되는 경우 굳이 DTO 객체를 생성하지 않도록 서비스 메서드 변경.
- 변경사항이 반영된 테스트 코드 재작성.

## Test
- 한글로 @DisplayName 어노테이션을 사용하여 가독성 있는 테스트 코드 작성.
- 테스트 코드의 이름을 좀 더 적절하게 변경(ControllerTest -> ServiceTest 등).
- 테스트용 프로필(application-test.properties) 적용.

## DTO
- BoardArticleDTO의 게시글 작성자 Id 필드를 해당 게시글이 로그인된 사용자에 의해 작성되었는지 여부를 나타내는 필드로 변경.

## etc
- LoginSessionInfo는 대부분 BoardAccount 엔티티를 기반으로 생성되기 때문에 해당 생성자 구현.
